requirements.txt
# Web scraping
playwright==1.41.0
beautifulsoup4==4.12.3

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Scheduling
apscheduler==3.10.4

# WhatsApp notifications
twilio==8.13.0

# AI parsing
openai==1.12.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.1

# Utilities
httpx==0.26.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.24.2

# Logging
structlog==24.1.0
